+++ Editor/Package.cs
@@ -5,7 +5,7 @@ namespace Hananoki.CustomHierarchy {
   public static class Package {
     public const string name = "CustomHierarchy";
     public const string editorPrefName = "Hananoki.CustomHierarchy";
-    public const string version = "0.6.5-preview";
+    public const string version = "0.6.6-preview";
   }
   
 #if UNITY_EDITOR